Move control of the cursor icon to WlDataDevice
