+++ src/server/frontend_wayland/wl_data_source.h
@@ -18,7 +18,6 @@
 #define MIR_FRONTEND_WL_DATA_SOURCE_H_
 
 #include "wayland_wrapper.h"
-#include "wl_surface.h"
 
 namespace mir
 {
@@ -31,22 +30,19 @@ class DataExchangeSource;
 
 namespace frontend
 {
-class DragIconController;
-
 class WlDataSource : public wayland::DataSource
 {
 public:
     WlDataSource(
         wl_resource* new_resource,
         std::shared_ptr<Executor> const& wayland_executor,
-        scene::Clipboard& clipboard,
-        std::shared_ptr<DragIconController> drag_icon_controller);
+        scene::Clipboard& clipboard);
     ~WlDataSource();
 
     static auto from(struct wl_resource* resource) -> WlDataSource*;
 
     void set_clipboard_paste_source();
-    void start_drag_n_drop_gesture(std::optional<wl_resource*> const& icon);
+    void start_drag_n_drop_gesture();
 
     /// Wayland requests
     /// @{
@@ -55,19 +51,6 @@ public:
     /// @}
 
 private:
-    class DragIconSurface : public NullWlSurfaceRole
-    {
-    public:
-        DragIconSurface(WlSurface* icon, std::shared_ptr<DragIconController> drag_icon_controller);
-        ~DragIconSurface();
-
-        auto scene_surface() const -> std::optional<std::shared_ptr<scene::Surface>> override;
-
-    private:
-        wayland::Weak<WlSurface> const surface;
-        std::shared_ptr<scene::Surface> shared_scene_surface;
-        std::shared_ptr<DragIconController> const drag_icon_controller;
-    };
     class ClipboardObserver;
     class Source;
 
@@ -80,7 +63,6 @@ private:
     std::shared_ptr<Executor> const wayland_executor;
     scene::Clipboard& clipboard;
     std::shared_ptr<ClipboardObserver> const clipboard_observer;
-    std::shared_ptr<DragIconController> const drag_icon_controller;
     std::vector<std::string> mime_types;
     std::weak_ptr<scene::DataExchangeSource> paste_source;
     bool clipboards_paste_source_is_ours{false};
@@ -89,7 +71,6 @@ private:
     bool dnd_source_source_is_ours{false};
     uint32_t dnd_actions;
     std::optional<uint32_t> dnd_action;
-    std::optional<DragIconSurface> drag_surface;
 };
 }
 }
